chore: add current URL manipulation for Spotify callback
The `handleSpotifyCallback` function now includes the retrieval and manipulation of the current URL to dynamically form the `redirectUri`. A new `const currentUrl` retrieves the window location, and `redirectUri` is revised accordingly.

chore: modify URL handling in Spotify login redirection

The `redirectToSpotifyLogin` function now dynamically constructs the `redirectUri` using the current URL, enhancing the flexibility of the redirection process to Spotify. A new `const currentUrl` is added to handle this.

chore: remove unused constants in PlaylistImport component

The `clientId` and `redirectUri` constants, which were unnecessary in the `PlaylistImport` component, have been removed to clean up the code.